Initial commit of servo_lib.c
Init function made for OC2 to be used to control the servo. Uses timer 3 to create a 20ms period. The set compare function will need to be used to test the range of motion for our servo. A guess based on another servo is the val 2000 (1ms) is all the way to the left and the val 4000 (2ms) is all the way to the right,